clang-tidy: fix bugprone-*
This commit aims to fix all bugprone-* warnings from clang-tidy.
Here is the list of the ones found in project:
  "bugprone-assignment-in-if-condition",
  "bugprone-branch-clone",
  "bugprone-crtp-constructor-accessibility",
  "bugprone-empty-catch",
  "bugprone-forward-declaration-namespace",
  "bugprone-implicit-widening-of-multiplication-result",
  "bugprone-macro-parentheses",
  "bugprone-narrowing-conversions",
  "bugprone-parent-virtual-call",
  "bugprone-reserved-identifier",
  "bugprone-return-const-ref-from-parameter",
  "bugprone-suspicious-include",
  "bugprone-switch-missing-default-case",
  "bugprone-unchecked-optional-access"

+++ launcher/minecraft/auth/MinecraftAccount.cpp
@@ -49,6 +49,7 @@
 #include <QDebug>
 
 #include <QPainter>
+#include <cstddef>
 
 #include "minecraft/auth/AccountData.h"
 #include "minecraft/auth/AuthFlow.h"
@@ -231,9 +232,9 @@ bool MinecraftAccount::shouldRefresh() const
     auto expiresTimestamp = data.yggdrasilToken.notAfter;
 
     if (!expiresTimestamp.isValid()) {
-        expiresTimestamp = issuedTimestamp.addSecs(24 * 3600);
+        expiresTimestamp = issuedTimestamp.addSecs(static_cast<qint64>(24 * 3600));
     }
-    if (now.secsTo(expiresTimestamp) < (12 * 3600)) {
+    if (now.secsTo(expiresTimestamp) < (static_cast<qint64>(12 * 3600))) {
         return true;
     }
     return false;
